Update Queryhelper.cfc, resolves odd column names
Change the column name creation in queryhelper.cfc to fix an issue with specifically the aging report, but could show up elsewhere if illegal characters were in the column name. Code also appends custom_ in front of custom fields, and appends just an _ if the field starts with a numeric character, which is illegal only as the starting character in Coldfusion Query objects.